src_cloud_compute/requirements.txt
google-cloud-aiplatform
numpy
torch